+++ build_runner/test/io/build_output_reader_test.dart
@@ -14,6 +14,7 @@ import 'package:build_runner/src/build/build_state/build_step_id.dart';
 import 'package:build_runner/src/build/build_state/build_step_result.dart';
 import 'package:build_runner/src/build/build_state/post_process_build_step_id.dart';
 import 'package:build_runner/src/build/build_state/post_process_build_step_result.dart';
+import 'package:build_runner/src/build/shared_part_accumulator.dart';
 import 'package:build_runner/src/build_plan/build_directory.dart';
 import 'package:build_runner/src/build_plan/build_filter.dart';
 import 'package:build_runner/src/build_plan/build_options.dart';
@@ -165,6 +166,50 @@ void main() {
       );
       expect(await reader.canRead(notDeletedId), true);
       expect(await reader.canRead(deletedId), false);
+    });
+
+    test('can fully read generated shared parts (retaining imports)', () async {
+      final inputId = AssetId('a', 'lib/a.dart');
+      final expectedGeneratedPartId = AssetId('a', 'lib/_br_/a.dart');
+
+      final buildStepId = BuildStepId(primaryInput: inputId, phaseNumber: 0);
+      buildState.addBuildStepResult(
+        step: buildStepId,
+        result: BuildStepResult((b) {
+          b.result = true;
+          b.inArtifactTree = false;
+          b.wrotePartContribution = true;
+        }),
+      );
+      buildState.addSharedPart(SharedPartAccumulator(inputId, '// @dart=3.0'));
+      buildState.updatePartContribution(inputId, 0, [
+        'import \'package:foo/foo.dart\';',
+      ], '// contribution');
+
+      final buildPlan = await BuildPlan.load(
+        await BuildSpec.load(
+          builderFactories: BuilderFactories({}),
+          buildOptions: BuildOptions.forTests(),
+          testingOverrides: TestingOverrides(
+            buildPhases: buildPhases,
+            readerWriter: readerWriter,
+            buildPackages: buildPackages,
+          ),
+        ),
+      );
+      reader = BuildOutputReader(
+        buildPackages: buildPlan.buildSpec.buildPackages,
+        readerWriter: buildPlan.readerWriter,
+        buildState: buildState.toFinishedBuildState(),
+      );
+
+      expect(await reader.canRead(expectedGeneratedPartId), true);
+      final readResult = await reader.read(expectedGeneratedPartId);
+      expect(readResult.canRead, true);
+      final content = readResult.readAsString();
+      expect(content, contains('import \'package:foo/foo.dart\';'));
+      expect(content, contains('// @dart=3.0'));
+      expect(content, contains('// contribution'));
     });
 
     test('Failed steps interact well with build filters ', () async {